+++ Changelog.md
@@ -14,6 +14,8 @@ Released on ?.
 - Improved resolution of location pin images
 - Fixed Alptherm auto login
 - Fixed downloading images from weather pages that need a login
+- Disable swiping in track details when hovering over the track's height
+  profile
 
 ## Version 1.14.0 "New Zealand 2022/2023" edition